requirements.txt
torch==2.2.0
torchvision==0.17.0
transformers
Pillow==9.4.0
pytest==7.3.0
timm